requirements.txt
django>=2.0,<3.0
psycopg2>=2.7,<3.0
whitenoise
gunicorn
django_heroku
Pillow
requests
